pyproject.toml
[tool.poetry]
name = "tinkerbot"
version = "0.1.0"
description = "TinkerHub Event mannagement bot"
authors = ["winter_x64 <75623356+winter-x64@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nextcord = "^2.6.0"
firebase-admin = "^6.2.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
